seperate files and clean up code
